Add ABOUT submenu and credits to Port Menu
- Add ABOUT button linking to detailed about submenu
- Add credits at bottom of Port Menu showing:
  - C++ Port By Rich Vickers
  - GitHub repo link
  - Based on Zlatkok's original port
  - Original port GitHub link
- Include about menu with full credits and version info

+++ src/game/gameTime.cpp
@@ -132,7 +132,7 @@ static void initTimeDelta()
 
 static bool isGameAtMinute(dword minute)
 {
-    return minute == m_gameTimeInMinutes;
+    return static_cast<int>(minute) == m_gameTimeInMinutes;
 }
 
 static void endFirstHalf()
